requirements.txt
openai
python-dotenv
google-genai
numpy
pandas
tqdm
pillow

#matplotlib
#plotly
#scikit-learn
#pip install google-genai numpy
